+++ src/components/mainContent/MainContent.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import {MainContentContainer} from './MainContent.styles';
-import CardsView from './CardsView';
+import CardsView from '../cardsView/CardsView';
 
 interface MainContentProps {
     isSidebarOpen: boolean;
@@ -12,14 +12,9 @@ const MainContent: React.FC<MainContentProps> = ({isSidebarOpen, selectedDeck}) 
 
     return (
         <MainContentContainer isSidebarOpen={isSidebarOpen}>
-            {selectedDeck != null && (
-                <CardsView selectedDeck={selectedDeck} />
-            )}
-            {selectedDeck == null && (
-                <div>
-                    <h2>{'Select a Deck'}</h2>
-                </div>
-            )}
+            {selectedDeck != null
+                ? (<CardsView selectedDeck={selectedDeck}/>)
+                : (<h2>Select a Deck</h2>)}
         </MainContentContainer>
     );
 };
